+++ test/tabs_area_theme_data_test.dart
@@ -4,41 +4,6 @@ import 'package:tabbed_view/src/tabbed_view_icons.dart';
 import 'package:tabbed_view/src/theme/tabs_area_theme_data.dart';
 
 void main() {
-  group('close icon', () {
-    test('default', () {
-      TabsAreaThemeData data = TabsAreaThemeData();
-      expect(data.closeIconData, isNull);
-      expect(data.closeIconPath, isNotNull);
-    });
-
-    test('null', () {
-      TabsAreaThemeData data =
-          TabsAreaThemeData(closeIconData: null, closeIconPath: null);
-      expect(data.closeIconData, isNull);
-      expect(data.closeIconPath, isNotNull);
-    });
-
-    test('path', () {
-      TabsAreaThemeData data =
-          TabsAreaThemeData(closeIconPath: TabbedViewIcons.close);
-      expect(data.closeIconData, isNull);
-      expect(data.closeIconPath, isNotNull);
-    });
-
-    test('data', () {
-      TabsAreaThemeData data = TabsAreaThemeData(closeIconData: Icons.close);
-      expect(data.closeIconData, isNotNull);
-      expect(data.closeIconPath, isNull);
-    });
-
-    test('both', () {
-      TabsAreaThemeData data = TabsAreaThemeData(
-          closeIconData: Icons.close, closeIconPath: TabbedViewIcons.close);
-      expect(data.closeIconData, isNotNull);
-      expect(data.closeIconPath, isNotNull);
-    });
-  });
-
   group('menu icon', () {
     test('default', () {
       TabsAreaThemeData data = TabsAreaThemeData();